extract shared logic
